Fixed mirroring issue - update hinge angles have to be negated
